requirements.txt
pywhispercpp>=1.2.0
ffmpeg-python>=0.2.0
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
